+++ news/story.go
@@ -8,16 +8,16 @@ import (
 )
 
 type Story struct {
-	Id       int       `json:"id"`
+	Id       int64     `json:"id"`
 	Title    string    `json:"title"`
 	Url      string    `json:"url"`
-	Score    int       `json:"score"`
+	Score    int64     `json:"score"`
 	By       string    `json:"by"`
 	Time     Timestamp `json:"time"`
 	Comments []int     `json:"kids"`
 }
 
-func GetStory(id int) (*Story, error) {
+func GetStory(id int64) (*Story, error) {
 	resp, err := http.Get(fmt.Sprintf("%s/item/%d.json", baseUrl, id))
 	if err != nil {
 		return nil, err
